pruebas90000000

+++ utils/src/utils/pcb.c
@@ -2,7 +2,7 @@
 
 u_int32_t pid_counter = 0;
 
-t_pcb* inicializar_pcb(u_int32_t pid, char* archivo_pseudocodigo, u_int32_t tamanio_proceso, u_int32_t estimacion_inicial) {
+t_pcb* inicializar_pcb(u_int32_t pid, char* archivo_pseudocodigo, u_int32_t tamanio_proceso, double estimacion_inicial) {
     t_pcb* pcb = malloc(sizeof(t_pcb));
     pcb->pid = pid;
     pcb->archivo_pseudocodigo = strdup(archivo_pseudocodigo);
